pyproject.toml
[project]
name = "TechTrand"
version = "0.1.0"
description = "A Python project using uv for dependency management."
authors = [{name = "Jongkook Kim", email = "gaulbi@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "trafilatura",
    "requests",
    "pyyaml",
    "python-dotenv",
    "openai",
    "chromadb", 
    "sentence-transformers",
    "python-dotenv",
    "anthropic",
    "beautifulsoup4",
    "lxml",
    "urllib3",
    "mypy",
    "html2text",
    "typing-extensions",
    "voyageai",
    "google-generativeai",
    "sentence-transformers",
    "tenacity",
    "langchain",
    "langchain-text-splitters",
    "feedparser",
    "aiohttp",
    "pydantic",
    "httpx",
    "readability-lxml"
]

[tool.uv]
# Add uv-specific configuration here if needed
